Importing all files from current Trail Blazer.
This commit uploads all of the development files that are in the development build of Trail Blazer. There are some differences in this build compared to the build you can find on itch.io that I will have to remove in order for it to be the same as the current 1.3.1 release.